requirements.txt
# --- Web Framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# --- Database & ORM ---
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# --- Data Validation & Settings ---
pydantic>=2.6.0
pydantic-settings>=2.2.1

# --- Security (used in later phases, safe to include now) ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# --- Utilities ---
python-dotenv>=1.0.1

# --- Testing ---
pytest>=8.0.0
httpx>=0.27.0
